Implement first, last, init and tail for Array
Simple functions for slicing arrays. A work in progress as all tests are
currently implemented as smoke tests instead of actual spec tests -- due to a
lack of tooling.

The `init` and `tail` methods always return Arrays. The `first` and `last`
methods are a bit odd in that they can return single elements _or_ arrays
depending on the number requested. Not sure whether to go with pragmatism or
purity in this situation...